backend/requirements.txt
click==8.1.3
Flask==2.2.2
gunicorn==20.1.0
grpcio==1.51.1
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
Werkzeug==2.2.2
zipp==3.8.1
opencv-python-headless==4.6.0.066
protobuf==4.21.12